This commit finishes off this assignment!
Adds the complete Jasmine testing suite, which thoroughly tests the global dashboard controller and the widgetGraph directive.

Adds jasmine via the jasmine rails gem, which is really useful for structuring everything. No need to write the code to run the tests, since it comes with a built-in jasmine server that automatically runs the specs if you have the spec files in the right location.

I feel like this does pretty well for code coverage, and actually including meaningful tests in the suite.

Also, this commit separates out the widgetGraphTemplate code into its own template file such that I can include it separately from the jasmine test suite. This is so that I am able to really test the widgetGraph directive.

While generally testing went off without a hitch, I had a hard time getting the directive template to correctly become available in jasmine scope for testing.